Separate out configuration renders
