+++ dashboards-observability/public/components/application_analytics/components/create.tsx
@@ -3,13 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { 
-  EuiAccordion,
-  EuiBadge,
+import {
   EuiBasicTable,
   EuiButton,
-  EuiComboBox,
-  EuiComboBoxOptionOption,
   EuiFieldText, 
   EuiFlexGroup, 
   EuiFlexItem, 
@@ -27,93 +23,30 @@ import {
   EuiPopover, 
   EuiSelect, 
   EuiSpacer, 
-  EuiTableFieldDataColumnType, 
-  EuiText, 
+  EuiTableFieldDataColumnType,
   EuiTitle 
 } from "@elastic/eui";
-import { Autocomplete } from "../../common/search/autocomplete";
-import DSLService from "public/services/requests/dsl";
+import DSLService from "public/services/requests/dsl";  
 import React, { useEffect, useState } from "react";
 import { ChangeEvent } from "react";
-import { PPLReferenceFlyout } from "../../common/helpers";
-import { uiSettingsService } from "../../../../common/utils";
 import { AppAnalyticsComponentDeps } from "../home";
-import { ServiceMap } from "../../../components/trace_analytics/components/services";
-import { handleServiceMapRequest } from "../../../components/trace_analytics/requests/services_request_handler";
-import { ServiceObject } from "../../../components/trace_analytics/components/common/plots/service_map";
-import { FilterType } from "../../../components/trace_analytics/components/common/filters/filters";
 import { TraceConfig } from './trace_config';
+import { ServiceConfig } from "./service_config";
+import { LogConfig } from "./log_config";
+import { PPLReferenceFlyout } from "../../../components/common/helpers";
 
 interface CreateAppProps extends AppAnalyticsComponentDeps {
   dslService: DSLService;
 };
 
 export const CreateApp = (props: CreateAppProps) => {
-  const { parentBreadcrumb, chrome, dslService, query, setQuery, filters, setFilters, http } = props;
+  const { parentBreadcrumb, chrome } = props;
+  const [isPopoverOpen, setIsPopoverOpen] = useState(false);
+  const [isFlyoutVisible, setIsFlyoutVisible] = useState(false);
   const [state, setState] = useState({
     name: '',
     description: ''
   });
-  const [logOpen, setLogOpen] = useState(false);
-  const [servicesOpen, setServicesOpen] = useState(false);
-  const [traceOpen, setTraceOpen] = useState(false);
-  const [isPopoverOpen, setIsPopoverOpen] = useState(false);
-  const [isFlyoutVisible, setIsFlyoutVisible] = useState(false);
-  const [tempQuery, setTempQuery] = useState<string>('');
-  const [serviceMap, setServiceMap] = useState<ServiceObject>({});
-  const [serviceMapIdSelected, setServiceMapIdSelected] = useState<'latency' | 'error_rate' | 'throughput'>('latency');
-  const [selectedServices, setSelectedServices] = useState(filters.map((f) => { return { label: f.value }}));
-
-
-  useEffect (() => {
-    const serviceOptions = filters.filter(f => f.field === 'serviceName').map((f) => { return { label: f.value }});
-    const noDups = serviceOptions.filter((s, index) => { return serviceOptions.findIndex(ser => ser.label === s.label) === index });
-    setSelectedServices(noDups);
-  }, [filters])
-
-  const addFilter = (filter: FilterType) => {
-    for (const addedFilter of filters) {
-      if (
-        addedFilter.field === filter.field &&
-        addedFilter.operator === filter.operator &&
-        addedFilter.value === filter.value
-      ) {
-        return;
-      }
-    }
-    const newFilters = [...filters, filter];
-    setFilters(newFilters);
-  };
-
-  const onServiceChange = (selectedServices: any) => {
-    const serviceFilter = selectedServices.map((option: EuiComboBoxOptionOption<string>) => { 
-      return {
-        field: 'serviceName', 
-        operator: 'is', 
-        value: option.label, 
-        inverted: false, 
-        disabled: false 
-      }
-    })
-    setFilters(serviceFilter);
-  };
-
-  const services = Object.keys(serviceMap).map((service) => { return { label: service } });
-
-  const handleQueryChange = async (query: string) => setQuery(query);
-
-  const closeFlyout = () => {
-    setIsFlyoutVisible(false);
-  };
-
-  const showFlyout = () => {
-    setIsFlyoutVisible(true);
-  };
-
-  let flyout;
-  if (isFlyoutVisible) {
-    flyout = <PPLReferenceFlyout module="explorer" closeFlyout={closeFlyout} />;
-  }
 
   useEffect(() => {
     chrome.setBreadcrumbs(
@@ -128,8 +61,16 @@ export const CreateApp = (props: CreateAppProps) => {
       href: '#/application_analytics/create',
       },
     ]);
-    handleServiceMapRequest(http, dslService, serviceMap, setServiceMap);
     }, [])
+
+  const closeFlyout = () => {
+    setIsFlyoutVisible(false);
+  };
+
+    let flyout;
+    if (isFlyoutVisible) {
+      flyout = <PPLReferenceFlyout module="explorer" closeFlyout={closeFlyout} />;
+    }
 
   const dummyItems = [{id: '1', level: "Unavailable", color: "danger", conditions: "WHEN errorRate() IS ABOVE OR EQUAL TO 2%"}];
   const tableColumns = [
@@ -249,110 +190,11 @@ export const CreateApp = (props: CreateAppProps) => {
             </EuiPageContentHeaderSection>
           </EuiPageContentHeader>
           <EuiHorizontalRule />
-          <EuiAccordion
-            id="logSource"
-            buttonContent={
-              <>
-              <EuiText size="s">
-              <h3>Log Source</h3>
-              </EuiText>
-              <EuiSpacer size="s" />
-              <EuiText size="s" color="subdued">
-              Configure your application base query
-              </EuiText>
-              </>
-            }
-            extraAction={<EuiButton size="s" disabled={!logOpen} onClick={() => { handleQueryChange('') }}>Clear all</EuiButton>}
-            onToggle={(isOpen) => {setLogOpen(isOpen)}}
-            paddingSize="l"
-          >
-            <EuiFormRow
-            label="PPL Base Query"
-            helpText="The default logs view in the application will be filtered by this query."
-            >
-              <EuiFlexItem grow={false} key="query-bar" className="query-area">
-                <Autocomplete
-                  key={'autocomplete-bar'}
-                  query={query}
-                  tempQuery={tempQuery}
-                  handleQueryChange={handleQueryChange}
-                  handleQuerySearch={() => {}}
-                  dslService={dslService}
-                />
-                <EuiBadge 
-                  className={`ppl-link ${uiSettingsService.get('theme:darkMode') ? "ppl-link-dark" : "ppl-link-light"}`}
-                  color="hollow"
-                  onClick={() => showFlyout()}
-                  onClickAriaLabel={"pplLinkShowFlyout"}
-                >
-                  PPL
-                </EuiBadge>
-              </EuiFlexItem>
-            </EuiFormRow>
-          </EuiAccordion>
+          <LogConfig setIsFlyoutVisible={setIsFlyoutVisible} {...props} />
           <EuiHorizontalRule />
-          <EuiAccordion
-            id="servicesEntities"
-            buttonContent={
-              <>
-                <EuiText size="s">
-                <h3>
-                Services & Entities  <EuiBadge>{selectedServices.length}</EuiBadge>
-                </h3>
-              </EuiText>
-              <EuiSpacer size="s" />
-              <EuiText size="s" color="subdued">
-                Select services & entities to include in this application
-              </EuiText>
-              </>
-              }
-            extraAction={<EuiButton size="s" disabled={!servicesOpen} onClick={() => {setFilters([])}}>Clear all</EuiButton>}
-            onToggle={(isOpen) => {setServicesOpen(isOpen)}}
-            paddingSize="l"
-          >
-            <EuiFormRow
-            label="Services & Entities"
-            >
-              <EuiComboBox
-                aria-label="Select services and entities"
-                placeholder="Select services and entities"
-                options={services}
-                selectedOptions={selectedServices}
-                onChange={onServiceChange}
-                isClearable={false}
-                data-test-subj="servicesEntitiesComboBox"
-              />
-            </EuiFormRow>
-            <EuiSpacer />
-            <ServiceMap
-              serviceMap={serviceMap}
-              idSelected={serviceMapIdSelected}
-              setIdSelected={setServiceMapIdSelected}
-              addFilter={addFilter}
-            />
-          </EuiAccordion>
+           <ServiceConfig {...props} />
           <EuiHorizontalRule />
-          <EuiAccordion
-            id="traceGroups"
-            buttonContent={
-              <>
-                <EuiText size="s">
-                <h3>
-                Trace Groups  <EuiBadge>0</EuiBadge>
-                </h3>
-                </EuiText>
-                <EuiSpacer size="s" />
-                <EuiText size="s" color="subdued">
-                  Constrain your application to specific trace groups
-                </EuiText>
-              </>
-              }
-            extraAction={<EuiButton size="s" disabled={!traceOpen}>Clear all</EuiButton>}
-            onToggle={(isOpen) => {setTraceOpen(isOpen)}}
-            paddingSize="l"
-          >
             <TraceConfig {...props}/>
-          </EuiAccordion>
         </EuiPageContent>
         <EuiSpacer />
         <EuiPageContent id="availability">
